requirements.txt
paddleocr==2.8.1
paddleocr.egg==info
Pillow==11.0.0
